change include path
This commit undoes the change introduced in cd3137edf1975e1feb0ec3f719b4db4808dafd2e

From ktd3rd:

The reason I got rid of the aces/ part of that is so that local cmake compiles where I don't do an install but use the cmake config mechanism to find my compiled folder, the header file is not yet placed in the install location (aces/), and is just in the local folder. So I made the installed rules include flags such that the include path is added "-I<prefix>/aces", such that #include <aces_Writer.h> works, which matches what the include flags are for OpenEXR (i.e. with the -I) when installed.