Update build system (#115)
* fix some installation paths

Use GNUInstallDirs variable throughout the build system.

Bug: https://github.com/ampas/CTL/issues/100
Signed-off-by: Bernd Waibel <waebbl-gentoo@posteo.net>

* streamline finding OpenEXR and Imath / IlmBase

The search for OpenEXR and IlmBase resp. Imath dependencies can be
simplified, as OpenEXR depends on either Imath or IlmBase, depending
if we use <OpenEXR-3 or >=OpenEXR-3.

Signed-off-by: Bernd Waibel <waebbl-gentoo@posteo.net>

* enable testing only when CMAKE_BUILD_TESTS=ON

The function enable_testing was globally enabled, no matter the value
of the CMAKE_BUILD_TESTS option. This patch guards this call, so it
only gets called, when CMAKE_BUILD_TESTS=ON.

Signed-off-by: Bernd Waibel <waebbl-gentoo@posteo.net>

* require C++ 11 using set(CMAKE_CXX_STANDARD 11)

* add #include <limits> for std::numeric_limits

* remove pkg-config *.pc.in and cmake *.cmake.in files

Signed-off-by: Bernd Waibel <waebbl-gentoo@posteo.net>
Co-authored-by: michaeldsmith <37905569+michaeldsmith@users.noreply.github.com>